made the admin page CSS, updated the features, everything is ok
