+++ public/assets/css/adminInviteUser.scss
@@ -2,15 +2,16 @@
     background-color: var(--rosePastel);
     border-radius: 10px;
     box-shadow: 1px 1px 5px gray;
-    display: flex;
-    flex-direction: column;
-    width: 90%;
+    width: 95vw;
+    height:270px;
     margin: auto;
+    margin-bottom:8px;
 }
 
 .title-invite-user {
     padding: 8px;
     font-family: var(--fontTitle);
+    font-size: clamp(20px, 2em, 2.8vw);
     align-items: center;
     text-align: center;
     margin: 0;
@@ -21,6 +22,10 @@
     margin: auto;
     align-items: center;
     text-align: center;
+    height:216px;
+    display:flex;
+    flex-direction: column;
+    justify-content: space-between;
 }
 
 .invite-user-errors {
@@ -44,25 +49,82 @@
 
 .label-form-email {
     font-family: var(--fontText);
-    display: block;
     text-align: center;
     padding: 8px;
 }
 
-#invitedEmail {
-    margin: 8px;
-    line-height: 24px;
+
+.input-container {
+    height: 50px;
+    position: relative;
+    width: 100%;
+}
+
+.input-invite {
+    background-color: var(--bleu);
+    border-radius: 12px;
+    border: 0;
+    box-sizing: border-box;
+    font-size: 18px;
+    height: 100%;
+    outline: 0;
+    padding: 4px 20px 0;
+    width: 100%;
+}
+
+.cut {
+    background-color: var(--bleuVert);
+    border-radius: 10px;
+    height: 20px;
+    left: 20px;
+    position: absolute;
+    top: -20px;
+    transform: translateY(0);
+    transition: transform 200ms;
+    width: 50px;
+}
+
+
+.input-invite:focus ~ .cut,
+.input-invite:not(:placeholder-shown) ~ .cut {
+    transform: translateY(10px);
+
+}
+.invite-placeholder {
+    color: var(--saumon);
+    font-family: var(--fontText);
+    left: 20px;
+    line-height: 14px;
+    pointer-events: none;
+    position: absolute;
+    transform-origin: 0 50%;
+    transition: transform 200ms, color 200ms;
+    top: 20px;
+}
+
+.input-invite:focus ~ .invite-placeholder,
+.input-invite:not(:placeholder-shown) ~ .invite-placeholder {
+    transform: translateY(-25px) translateX(10px) scale(0.75);
+}
+
+.input-invite:not(:placeholder-shown) ~ .invite-placeholder {
+    color: var(--saumon);
+}
+
+.input-invite:focus ~ .invite-placeholder {
+    color: var(--saumon);
 }
 
 .cta {
     position: relative;
-    margin: auto;
+    //margin: auto;
     padding: 20px;
     transition: all 0.2s ease;
     background-color: var(--rosePastel);
     border:0;
     text-align: center;
     margin-top:16px;
+    margin-bottom:8px;
 }
 .cta:before {
     content: "";
@@ -87,7 +149,7 @@
 }
 .cta svg {
     position: relative;
-    top: -4px;
+    top: 2px;
     margin-left: 10px;
     fill: none;
     stroke-linecap: round;
