+++ public/assets/css/userHeader.css
@@ -1,5 +1,5 @@
 .mobile-header-container {
-  margin-bottom: 64px;
+  margin-bottom: 8px;
   position: relative;
   height: 315px;
   background-color: var(--bleuVert);
