+++ public/assets/css/userHeaderDesktop.css
@@ -7,7 +7,8 @@
   position: relative;
   background-color: var(--bleuVert);
   padding-top: 48px;
-  padding-bottom: 48px;
+  padding-bottom: 32px;
+  margin-bottom: 8px;
 }
 
 .avatar-container-desktop {
@@ -163,6 +164,44 @@
 .img-header-deco.avatar6 {
   bottom: 7%;
   right: -20vw;
+}
+
+.user-gestion-link-desktop {
+  display: block;
+  margin: auto;
+  width: max-content;
+  position: relative;
+  margin-top: -8px;
+}
+.user-gestion-link-desktop:before {
+  content: attr(data-hover);
+  visibility: hidden;
+  opacity: 0;
+  width: max-content;
+  background-color: var(--smallButtonColor);
+  color: #fff;
+  text-align: center;
+  border-radius: 5px;
+  padding: 5px 5px;
+  transition: opacity 1s ease-in-out;
+  position: absolute;
+  z-index: 1;
+  top: 90%;
+  left: -3%;
+}
+.user-gestion-link-desktop:hover:before {
+  opacity: 1;
+  visibility: visible;
+}
+
+.img-gestion-desktop {
+  width: 100px;
+  height: 100px;
+  transition: width, 6s, height, 0.6s;
+}
+.img-gestion-desktop:hover {
+  width: 120px;
+  height: 120px;
 }
 
 /*# sourceMappingURL=userHeaderDesktop.css.map */
