+++ public/assets/css/userPlanning.css
@@ -3,6 +3,7 @@
   width: 95%;
   padding: 16px;
   margin: auto;
+  margin-bottom: 8px;
   border-radius: 20px;
 }
