+++ src/View/Admin/delete.html.twig
@@ -3,28 +3,30 @@
     <link rel="stylesheet" href="/assets/css/style.css">
 {% endblock %}
 {% block content %}
-    <div class="container">
-        <div class="form">
-            <h2 class="title-deleteUSer">Supprimer un utilisateur</h2>
-            <form action='delete' method='post' class="deleteUser">
-                <div class="deleteformerror">
-                    {% for messages in message %}
-                        <li>{{ messages }}</li>
-                    {% endfor %}
-                </div>
-                <label for="email"> Email: </label>
-                <input type='email' id='email' name='email'>
-                <button type='submit'> Supprimer</button>
-            </form>
-        </div>
+    <div class="delete-container">
+        <h2 class="title-deleteUSer">Supprimer un utilisateur</h2>
+        <form action='delete' method='post' class="deleteUser">
+
+                {% for messages in message %}
+                    <li class="deleteformerror">{{ messages }}</li>
+                {% endfor %}
+
+                <p class="label-delete"> Veuillez entrer le mail de la personne que vous souhaitez supprimer </p>
+            <div class="input-container login-email-container">
+                <input id="email" name="email" class="input-delete" type="email" placeholder=" " required/>
+                <div class="cut"></div>
+                <label for="email" class="placeholder">Email</label>
+            </div>
+            <button class="cta">
+                <span>Supprimer l'utilisateur</span>
+                <svg width="13px" height="10px" viewBox="0 0 13 10">
+                    <path d="M1,5 L11,5"></path>
+                    <polyline points="8 1 12 5 8 9"></polyline>
+                </svg>
+            </button>
+        </form>
     </div>
 {% endblock %}
-
-
-
-
-
-
