+++ src/View/Admin/index.html.twig
@@ -1,7 +1,6 @@
 {% extends 'layout.html.twig' %}
 {% block stylesheet %}
-    <link rel="stylesheet" href="/assets/css/adminInviteUser.css">
-    <link rel="stylesheet" href="/assets/css/delete.css">
+    <link rel="stylesheet" href="/assets/css/adminPage.css">
 {% endblock %}
 {% block title %}GESTION{% endblock %}
 {% block header %}
@@ -9,7 +8,10 @@
     {% include 'userHeaderDesktop.html.twig' %}
 {% endblock %}
     {% block content %}
-        {% include 'Admin/adminInviteUser.html.twig' %}
+        <div class="users-gestion">
+            {% include 'Admin/adminInviteUser.html.twig' %}
+            {% include 'Admin/delete.html.twig' %}
+        </div>
         {% include 'Admin/adminUserView.html.twig' %}
-        {% include 'Admin/delete.html.twig' %}
+
     {% endblock %}
